Extract OwnDatePicker
