+++ frontend/src/metabase/parameters/components/ParameterValuePicker/ParameterValuePicker.tsx
@@ -1,22 +1,10 @@
-import { useClickOutside } from "@mantine/hooks";
-import { useState } from "react";
-import { t } from "ttag";
 import _ from "underscore";
 
-import { DateAllOptionsWidget } from "metabase/components/DateAllOptionsWidget";
-import { DateMonthYearWidget } from "metabase/components/DateMonthYearWidget";
-import { DateQuarterYearWidget } from "metabase/components/DateQuarterYearWidget";
-import { DateRelativeWidget } from "metabase/components/DateRelativeWidget";
-import { formatParameterValue } from "metabase/parameters/utils/formatting";
 import { DefaultParameterValueWidget } from "metabase/query_builder/components/template_tags/TagEditorParamParts";
-import { Popover } from "metabase/ui";
 import { isDateParameter } from "metabase-lib/parameters/utils/parameter-type";
-import type { Parameter, ParameterType, TemplateTag } from "metabase-types/api";
+import type { Parameter, TemplateTag } from "metabase-types/api";
 
-import {
-  TextInputIcon,
-  TextInputTrirgger,
-} from "./ParameterValuePicker.styled";
+import { OwnDatePicker } from "./OwnDatePicker";
 import { PlainValueInput } from "./PlainValueInput";
 import { shouldShowPlainInput } from "./core";
 
@@ -29,20 +17,16 @@ interface ParameterValuePickerProps {
 }
 
 // TODO multiple value pickers
-// TODO make controlled outside
-// TODO must change value when type is changed
 // TODO setting default value on blur/closing picker
 // TODO error states
 // TODO placeholders unification
 // TODO filter input for numbers
+/**
+ * This component is designed to be controlled outside,
+ * without keeping its own state.
+ */
 export function ParameterValuePicker(props: ParameterValuePickerProps) {
-  const {
-    tag,
-    parameter,
-    value: initialValue,
-    onValueChange,
-    placeholder,
-  } = props;
+  const { tag, parameter, value, onValueChange, placeholder } = props;
 
   if (!parameter) {
     return null;
@@ -53,7 +37,7 @@ export function ParameterValuePicker(props: ParameterValuePickerProps) {
   if (shouldShowPlainInput(parameter)) {
     return (
       <PlainValueInput
-        value={initialValue}
+        value={value}
         onChange={onValueChange}
         placeholder={placeholder}
       />
@@ -63,7 +47,7 @@ export function ParameterValuePicker(props: ParameterValuePickerProps) {
   if (isDateParameter(parameter)) {
     return (
       <OwnDatePicker
-        value={initialValue}
+        value={value}
         parameter={parameter}
         onValueChange={onValueChange}
       />
@@ -73,8 +57,8 @@ export function ParameterValuePicker(props: ParameterValuePickerProps) {
   // The fallback
   return (
     <DefaultParameterValueWidget
-      parameter={getAmendedParameter(tag, parameter)}
-      value={initialValue}
+      parameter={DEPRECATED_getAmendedParameter(tag, parameter)}
+      value={value}
       setValue={onValueChange}
       isEditing
       commitImmediately
@@ -83,7 +67,10 @@ export function ParameterValuePicker(props: ParameterValuePickerProps) {
   );
 }
 
-function getAmendedParameter(tag: TemplateTag, parameter: Parameter) {
+function DEPRECATED_getAmendedParameter(
+  tag: TemplateTag,
+  parameter: Parameter,
+) {
   const amended =
     tag.type === "text" || tag.type === "dimension"
       ? parameter || {
@@ -101,103 +88,4 @@ function getAmendedParameter(tag: TemplateTag, parameter: Parameter) {
   // We want to remove "default" and "required" so that it
   // doesn't show up in the default value input icon
   return _.omit(amended, "default", "required");
-}
-
-// TODO popover z-index (select inside dropdown)
-function OwnDatePicker(props: {
-  value: any;
-  parameter: Parameter;
-  onValueChange: (value: any) => void;
-}) {
-  const { value, parameter, onValueChange } = props;
-  const [isOpen, setIsOpen] = useState(false);
-  const formatted = formatParameterValue(value, parameter);
-
-  const DateWidget = {
-    "date/relative": DateRelativeWidget,
-    "date/month-year": DateMonthYearWidget,
-    "date/quarter-year": DateQuarterYearWidget,
-    // pickers
-    "date/single": DateAllOptionsWidget,
-    "date/range": DateAllOptionsWidget,
-    "date/all-options": DateAllOptionsWidget,
-  }[parameter.type];
-
-  const openPopover = () => setIsOpen(true);
-  const closePopover = () => setIsOpen(false);
-
-  const [triggerRef, setTriggerRef] = useState<HTMLDivElement | null>(null);
-  const ref = useClickOutside(closePopover, null, [triggerRef]);
-
-  // console.log("OwnDatePicker", value, formatted);
-
-  const icon = value ? (
-    <TextInputIcon
-      name="close"
-      onClick={() => {
-        onValueChange(null);
-        setIsOpen(false);
-      }}
-    />
-  ) : (
-    <TextInputIcon name="chevrondown" />
-  );
-  // This is required to allow clicking through the "chevrondown" icon.
-  // Must be replaced with `rightSectionPointerEvents=none` after upgrade
-  const rightSectionProps = value
-    ? undefined
-    : { style: { pointerEvents: "none" } };
-
-  return (
-    <Popover opened={isOpen}>
-      <Popover.Target>
-        <TextInputTrirgger
-          ref={setTriggerRef}
-          value={typeof formatted === "string" ? formatted : value ?? ""} // required by Mantine
-          readOnly
-          placeholder={t`Select a default value…`}
-          onClick={openPopover}
-          rightSection={icon}
-          rightSectionProps={rightSectionProps}
-        />
-      </Popover.Target>
-
-      <Popover.Dropdown>
-        <div ref={ref}>
-          {DateWidget ? (
-            <DateWidget
-              value={value}
-              initialValue={getInitialDateValue(
-                value,
-                parameter.type as ParameterType,
-              )}
-              onClose={closePopover}
-              setValue={onValueChange}
-            />
-          ) : (
-            "<none>"
-          )}
-        </div>
-      </Popover.Dropdown>
-    </Popover>
-  );
-}
-
-function getInitialDateValue(value: any, parameterType: ParameterType) {
-  if (value == null) {
-    if (parameterType === "date/single") {
-      return getIsoDate();
-    }
-
-    if (parameterType === "date/range") {
-      const now = getIsoDate();
-      return `${now}~${now}`;
-    }
-  }
-
-  return value;
-}
-
-function getIsoDate() {
-  return new Date().toISOString().slice(0, 10);
 }
